refactor(backend): extract SQLiteReindexMixin from SQLiteDialect
Move REINDEX support methods (supports_reindex, supports_reindex_expressions,
format_reindex_statement) from SQLiteDialect class body to a dedicated
SQLiteReindexMixin, ensuring every SQLite-specific Protocol has a
corresponding Mixin implementation.
